requirements.txt
asgiref==3.2.7
autopep8==1.5.2
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coverage==5.1
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-allauth==0.41.0
django-appconf==1.0.3
django-autoslug==1.9.6
django-autoslug-field==0.2.3
django-avatar==5.0.0
django-bootstrap4==1.1.1
django-classy-tags==1.0.0
django-comments-dab==1.3.0
django-compat==1.0.15
django-contrib-comments==1.9.2
django-cookie-law==2.0.3
django-debug-toolbar==2.2
django-favicon-plus==0.0.8
django-froala-editor==3.1.0
django-imagekit==4.0.2
django-import-export==2.0.2
django-sesame==1.7
django-tinymce==3.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
et-xmlfile==1.0.1
Faker==4.0.3
idna==2.9
jdcal==1.4.1
lxml==4.5.0
MarkupPy==1.14
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pep8==1.7.1
pilkit==2.0
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pytest-runner==5.2
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
text-unidecode==1.3
ua-parser==0.10.0
urllib3==1.25.8
xlrd==1.2.0
xlwt==1.3.0
